pyproject.toml
[build-system]
requires = ["setuptools>=80.9.0", "wheel>=0.45.1"]
build-backend = "setuptools.build_meta"

[project]
name = "citation-verifier"
version = "0.1.0"
description = "A tool for verifying legal citations in briefs, memos, and other court filings"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Proprietary - See LICENSE"}
authors = [
    {name = "Phaethon Order LLC"}
]
keywords = ["legal", "citations", "verification", "court", "documents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Legal Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "build==1.3.0",
    "certifi==2025.8.3",
    "click==8.3.0",
    "courts-db==0.10.26",
    "eyecite==2.7.6",
    "fast-diff-match-patch==2.1.0",
    "fastapi==0.117.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "lxml==6.0.2",
    "numpy==2.3.3",
    "packaging==25.0",
    "pandas==2.3.2",
    "pip-tools==7.5.0",
    "psycopg==3.2.10",
    "psycopg-binary==3.2.10",
    "pyahocorasick==2.2.0",
    "pydantic==2.11.9",
    "pydantic-core==2.33.2",
    "pypdf==6.1.0",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.2.0",
    "python-dotenv==1.1.1",
    "python-jose[cryptography]==3.3.0",
    "pytz==2025.2",
    "regex==2025.9.18",
    "SQLAlchemy==2.0.36",
    "stripe==11.5.0",
    "reporters-db==3.2.58",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.48.0",
    "typing-inspection==0.4.1",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
    "uvicorn==0.36.0",
    "wheel==0.45.1",
]

[project.optional-dependencies]
dev = [
    "pip-tools==7.5.0",
]

[project.urls]
Homepage = "https://github.com/aonanj/citation-verifier"
Repository = "https://github.com/aonanj/citation-verifier"

[tool.setuptools]
packages = ["svc", "utils", "verifiers"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.txt"]

[tool.pip-tools]
# Configuration for pip-compile
generate-hashes = false

[tool.ruff]
line-length = 180
lint.select = ["E", "F", "W", "I"]  # enabled rules
lint.ignore = ["E203", "W293", "W291"]
exclude = ["build", "dist", ".venv"]
lint.extend-select = []
lint.per-file-ignores = { "tests/*" = ["F401"] }
